LayoutTests/imported/w3c/web-platform-tests/media-source/SourceBuffer-abort-updating-expected.txt


PASS SourceBuffer#abort() (video/webm; codecs="vorbis,vp8") : Check the algorithm when the updating attribute is true.
PASS SourceBuffer#abort() (video/mp4) : Check the algorithm when the updating attribute is true.
